refactor(client): ♻️ move avatar to lobby component
